refactor: extract voting tests
This extracts the previously nested `VotingTest`. `AppendResult` and
`StaticClusterMembershipService` become reusable too.
Smaller test helpers such as `awaitLeader` are just duplicated.
